update button in cell working fine

+++ SAB-Cars/SAB-Cars/CarsViewController.swift
@@ -61,10 +61,18 @@ extension CarsViewController:UITableViewDelegate,UITableViewDataSource {
         cell.carImg.image = UIImage(data: data!)
       if canDelet == true {
           cell.updateButton.isHidden = false
+          cell.updateButton.tag = indexPath.row
+          cell.updateButton.addTarget(self, action: #selector(buttonAction), for: .touchUpInside)
       }
-
         return cell
     }
+    
+    @objc func buttonAction(sender:UIButton){
+         let indexpath = IndexPath(row: sender.tag, section: 0)
+        let cell = cars[indexpath.row]
+         let next:AddCarViewController=self.storyboard?.instantiateViewController(withIdentifier: "next") as! AddCarViewController
+         self.navigationController?.pushViewController(next, animated: true)
+     }
     func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
         let indx = cars[indexPath.row].id
         let imageCar = cars[indexPath.row].carimg
@@ -74,17 +82,34 @@ extension CarsViewController:UITableViewDelegate,UITableViewDataSource {
         showvc.photo = imageCar
         navigationController?.pushViewController(showvc, animated: true)
     }
-    func tableView(_ tableView: UITableView, trailingSwipeActionsConfigurationForRowAt indexPath: IndexPath) -> UISwipeActionsConfiguration? {
-        
-        let actionDelete = UIContextualAction(style: .destructive, title: "Delete") { _, _, handler in
-            guard  self.canDelet == true else { return }
+//    func tableView(_ tableView: UITableView, trailingSwipeActionsConfigurationForRowAt indexPath: IndexPath) -> UISwipeActionsConfiguration? {
+//        
+//        let actionDelete = UIContextualAction(style: .destructive, title: "Delete") { _, _, handler in
+//            guard  self.canDelet == true else { return }
+//            let itemToDelete = self.cars[indexPath.row]
+//            self.dbStore.collection("Cars").document(itemToDelete.id).delete()
+//            self.db.child("Comments").child(itemToDelete.id).removeValue()
+//            self.getInfo()
+//            tableView.reloadData()
+//            }
+// 
+//        return UISwipeActionsConfiguration(actions: [actionDelete])
+//    }
+//    
+//    func tableView(_ tableView: UITableView, editingStyleForRowAt indexPath: IndexPath) -> UITableViewCell.EditingStyle {
+//        .delete
+//    }
+    func tableView(_ tableView: UITableView, commit editingStyle: UITableViewCell.EditingStyle, forRowAt indexPath: IndexPath) {
+        guard  self.canDelet == true else { return }
+        if editingStyle == .delete {
             let itemToDelete = self.cars[indexPath.row]
+            cars.remove(at: indexPath.row)
+            tableView.deleteRows(at: [indexPath], with: .automatic)
             self.dbStore.collection("Cars").document(itemToDelete.id).delete()
             self.db.child("Comments").child(itemToDelete.id).removeValue()
-//                self.fetchDataFromDB()
-            }
- 
-        return UISwipeActionsConfiguration(actions: [actionDelete])
+            self.getInfo()
+            tableView.reloadData()
+        }
     }
     
     func getInfo(){
@@ -93,6 +118,7 @@ extension CarsViewController:UITableViewDelegate,UITableViewDataSource {
             if let err = err {
                 print("Error getting documents: \(err)")
             } else {
+                self.cars.removeAll()
                 for document in querySnapshot!.documents {
                     // print("\(document.documentID) => \(document.data())")
                     let values = document.data()
